[rename] #20 오픈그래프 관련 라이브러리 external로 이동
