+++ src/main/java/org/sopt/sweet/global/external/opengraph/lib/OpenGraph.java
@@ -1,4 +1,4 @@
-package org.sopt.sweet.domain.opengraph.lib;
+package org.sopt.sweet.global.external.opengraph.lib;
 
 import org.htmlcleaner.HtmlCleaner;
 import org.htmlcleaner.TagNode;
